jsp 삭제, 회원정보로드버그 수정

+++ happyhouse-frontend/src/components/Admin/NewsManage.vue
@@ -106,7 +106,8 @@ export default {
       });
     },
     deleteNews(no) {
-      http.delete(`/api/news/${no}`).then(() => this.loadList());
+      if (confirm("해당 뉴스를 삭제하시겠습니까?"))
+        http.delete(`/api/news/${no}`).then(() => this.loadList());
     },
   },
 };
